test: reuse `EventType` enum for certain integration tests
Also guards certain integration tests with the required feature(s).
