+++ tests/common.rs
@@ -23,3 +23,20 @@ impl log::Log for Logger {
 pub fn logger_init() {
     let _ = log::set_logger(&Logger);
 }
+
+#[allow(dead_code, reason = "This is used by most tests but not all of them")]
+#[derive(Debug, PartialEq, Default)]
+pub enum EventType {
+    #[default]
+    Push,
+    PullRequest,
+}
+
+impl std::fmt::Display for EventType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Push => write!(f, "push"),
+            Self::PullRequest => write!(f, "pull_request"),
+        }
+    }
+}
